// FrontEnd_BackEnd_ex_ci_cd/react_cafe_ex/src/types/Product.ts
export interface Product { // 원래는 스프링의 Entity보고 작성해야하는 코드
    id: number;
    name: string;
    price: number;
    category: string;      // Enum은 JSON으로 오면 문자열
    stock: number;
    image: string;
    description: string;
    inputdate: string;     // LocalDate → 문자열로 넘어옴 (예: "2026-02-20")
}